Add custom 404 error page
The commit adds a new, customized 404 error page to replace the default error page. The new HTML file was created with defined features for better user experience such as specifying the character set, viewport, and page description for SEO purposes. On the other hand, a new script named 431-4810f262f984aefd.js has been added to handle the dynamic behavior of the page. This enhancement is important to provide users with more helpful error information and guide them to relative pages.