Update build script to prevent Jekyll from ignoring files
Modified the "build" command script in package.json to include "touch out/.nojekyll". This change was made to prevent Jekyll from ignoring output files that start with an underscore in our next.js build. This ensures that all of the necessary files are included when deploying our app on platforms that use Jekyll for rendering, such as GitHub Pages.